# Conflicts:
#	src/Ext/Building/Hooks.Grinding.cpp
#	src/Ext/Script/Body.cpp
#	src/Ext/Techno/Hooks.ReceiveDamage.cpp
#	src/Misc/Hooks.BugFixes.cpp
#	src/New/Entity/AttachEffectClass.cpp

+++ src/Ext/Techno/Hooks.cpp
@@ -1104,7 +1104,7 @@ DEFINE_HOOK(0x4DF3A6, FootClass_UpdateAttackMove_Follow, 0x6)
 	if (pTypeExt->AttackMove_Follow || pTypeExt->AttackMove_Follow_IfMindControlIsFull && pThis->CaptureManager && pThis->CaptureManager->CannotControlAnyMore())
 	{
 		auto const& pTechnoVectors = Helpers::Alex::getCellSpreadItems(pThis->GetCoords(),
-			pThis->GetGuardRange(2) / Unsorted::LeptonsPerCell, pTypeExt->AttackMove_Follow_IncludeAir);
+			pThis->GetGuardRange(2) / (double)Unsorted::LeptonsPerCell, pTypeExt->AttackMove_Follow_IncludeAir);
 
 		TechnoClass* pClosestTarget = nullptr;
 		int closestRange = 65536;
